requirements.txt
moviepy~=1.0.3
scikit-learn==1.4.1.post1
jupyter==1.0.0
nltk==3.8.1
wordcloud==1.9.3
librosa~=0.10.1
pandas~=2.2.1
numpy~=1.26.4
torch~=2.2.2
torchaudio~=2.2.2
seaborn~=0.13.2
stop-words==2018.7.23
xgboost==2.0.3
gensim==4.3.2
statsmodels==0.14.2
scipy==1.10.1
matplotlib~=3.9.0
mrmr-selection==0.2.8